webapp/llm_utils/requirements.txt
mkdocs
mkdocstrings
isort
streamlit
pandas
pyodbc
prompty
promptflow
langchain_core
langchain_community
langchain_openai
langchain_prompty
pypandoc
pymupdf
python-docx 
fastapi
python-magic
pydantic
pandas
docx